Speed up deserialization (#40)
* Add module 'ttlv'

This commit imports some draft optimization work intended for
'kmip-ttlv', with the goal of deprecating that crate entirely.  The fast
scanner and formatter defined here should provide good performance and
have a simple API, that 'src/de/mod.rs' needs to be rebased onto.

* [types] Derive 'Ordinalize' for all enums

* [types/common] Impl manual serde

The only breaking change is to 'CryptographicUsageMask', defining it via
'bitflags' instead of 'enum-flags'.  The latter hasn't been updated in a
while and 'bitflags' is the preferred crate for this purpose.

* [types/request] Impl manual serde

* [types/response] Impl manual serde (partially)

* Update request serialization tests to use the new Formatter.

---------

Co-authored-by: Ximon Eighteen <3304436+ximon18@users.noreply.github.com>